tests/wpt/web-platform-tests/html/canvas/offscreen/path-objects/2d.path.roundrect.radius.intersecting.2.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.path.roundrect.radius.intersecting.2</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.path.roundrect.radius.intersecting.2</h1>
<p class="desc">Check that roundRects with intersecting corner arcs are rendered correctly.</p>


<script>
var t = async_test("Check that roundRects with intersecting corner arcs are rendered correctly.");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

ctx.fillStyle = '#f00';
ctx.fillRect(0, 0, 100, 50);
ctx.roundRect(0, 0, 100, 50, [1000, 1000, 1000, 1000]);
ctx.fillStyle = '#0f0';
ctx.fill();
_assertPixel(offscreenCanvas, 2,25, 0,255,0,255, "2,25", "0,255,0,255");
_assertPixel(offscreenCanvas, 50,1, 0,255,0,255, "50,1", "0,255,0,255");
_assertPixel(offscreenCanvas, 50,25, 0,255,0,255, "50,25", "0,255,0,255");
_assertPixel(offscreenCanvas, 50,48, 0,255,0,255, "50,48", "0,255,0,255");
_assertPixel(offscreenCanvas, 97,25, 0,255,0,255, "97,25", "0,255,0,255");
_assertPixel(offscreenCanvas, 1,1, 255,0,0,255, "1,1", "255,0,0,255");
_assertPixel(offscreenCanvas, 98,1, 255,0,0,255, "98,1", "255,0,0,255");
_assertPixel(offscreenCanvas, 1,48, 255,0,0,255, "1,48", "255,0,0,255");
_assertPixel(offscreenCanvas, 98,48, 255,0,0,255, "98,48", "255,0,0,255");
t.done();

});
</script>
